Feat: Unit Test User

+++ src/users/user.service.ts
@@ -1,19 +1,47 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
+import { envVariableKeys } from 'src/common/const/env.const';
+import * as bcrypt from 'bcrypt'
+import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class UsersService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    private readonly configService: ConfigService,
   ){}
   
-  create(createUserDto: CreateUserDto) {
-    return this.userRepository.save(createUserDto);
+  async create(createUserDto: CreateUserDto) {
+    const { email, password } = createUserDto
+
+    const user = await this.userRepository.findOne({
+      where: {
+        email,
+      }
+    })
+
+    if(user){
+      throw new BadRequestException('이미 가입한 이메일 입니다.');
+    }
+
+    const hash = await bcrypt.hash(password, this.configService.get<number>(envVariableKeys.hashRound))
+    
+    await this.userRepository.save({
+      email,
+      password: hash,
+    })
+
+    return await this.userRepository.findOne({
+      where: {
+        email
+      }
+    })
+
   }
 
   findAll() {
@@ -35,6 +63,8 @@ export class UsersService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const {password} = updateUserDto
+
     const user = await this.userRepository.findOne({
       where: {
         id
@@ -44,10 +74,15 @@ export class UsersService {
     if (!user) {
       throw new NotFoundException('존재하지 않는 ID값의 유저입니다.')
     }
+
+    const hash = await bcrypt.hash(password, this.configService.get<number>(envVariableKeys.hashRound))
     
     await this.userRepository.update(
       {id},
-      updateUserDto,
+      {
+        ...updateUserDto,
+        password: hash,
+      },
     )
 
     return this.userRepository.findOne({
@@ -57,7 +92,19 @@ export class UsersService {
     });
   }
 
-  remove(id: number) {
-    return this.userRepository.delete(id);
+  async remove(id: number) {
+    const user = await this.userRepository.findOne({
+      where: {
+        id,
+      }
+    })
+
+    if(!user){
+      throw new NotFoundException('존재하지 않는 사용자입니다.')
+    }
+
+    await this.userRepository.delete(id)
+
+    return id
   }
 }
